requirements.txt
groq==0.22.0
aiofiles==23.2.1
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.3.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.1.0.post1
fastapi==0.104.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
Jinja2==3.1.2
MarkupSafe==3.0.2
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing-inspection==0.4.0
typing_extensions==4.13.1
uvicorn==0.24.0
